Move focus test to Tests_MarkupText
